store_backend/requirements.txt
certifi==2021.10.8
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
homebrew==0.2.5
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.21.4
pymongo==4.1.0
python-dotenv==0.20.0
scikit-learn==1.0.1
scipy==1.7.3
sklearn==0.0
threadpoolctl==3.0.0
Werkzeug==2.0.3
